# docs/bruno/fixtures-game-api/apispec/fixtures-game-api.openapi.yaml
...
openapi: 3.0.3
info:
  title: Fixtures Game Backend API
  version: 1.0.0
  description: API contract for the no-money football prediction game backend. Use
    `FE - ...` tags for frontend work and `Internal - ...` tags for backend jobs/admin
    maintenance.
servers:
- url: http://localhost:3001
  description: Local backend
...
paths:
...
  "/friends-groups/check-slug/{friendsGroupSlug}":
    get:
      tags:
      - FE - Groups And Invites
      summary: Check slug
      operationId: get_friends_groups_check_slug_friendsGroupSlug
      description: 'FE placement: create-group form validation. Purpose: checks whether
        the desired slug is available. Debounce this while typing after local lowercase
        kebab-case validation, and still handle POST /friends-groups 409 on submit
        because create remains the source of truth.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupSlug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SlugCheckResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups":
    post:
      tags:
      - FE - Groups And Invites
      summary: Create friends group with subscription
      operationId: post_friends_groups
      description: 'FE placement: submit button on create-friends-group screen. Purpose:
        creates an approved friends group immediately, activates the selected football
        subscription, adds creator as member, hydrates fixtures, and returns invite_token/joinLink.
        Notes: user can own only one group; accessType must be open or private.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateFriendsGroupRequest"
            example:
              payload:
                name: Local Test Group
                slug: "{{friendsGroupSlug}}"
                accessType: open
                subscription:
                  providerLeagueId: "{{providerLeagueId}}"
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateFriendsGroupResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/friends-groups/competitions":
    get:
      tags:
      - FE - Groups And Invites
      summary: List competition catalog
      operationId: get_friends_groups_competitions
      description: 'FE placement: create-friends-group league select. Purpose:
        lists real football competitions and current season metadata for display
        to. Backend self-syncs missing current season data from SportMonks once,
        then returns DB rows. Notes: send provider_league_id as providerLeagueId
        when creating the group; backend resolves the current season.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CompetitionCatalogResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/me/groups":
    get:
      tags:
      - FE - Groups And Invites
      summary: My groups
      operationId: get_friends_groups_me_groups
      description: 'FE placement: dashboard and group switcher. Purpose: lists friends
        groups current user belongs to, including invite_token and joinLink so the
        FE can rebuild share/invite UI after reload or cache clear. Also includes
        friends_group.league and friends_group.currentMatchweek so group cards can
        show labels like Premier League week 12. Includes score as the current user''s
        cumulative season/group score summary for each group.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MyGroupsResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookies set by /auth/sign-in or /auth/sign-up.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
      required:
      - error
...
    FriendsGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_by:
          type: string
          format: uuid
        invite_token:
          type: string
        is_open:
          type: boolean
        status:
          type: string
          enum:
          - approved
          - pending
          - rejected
          - archived
        review_note:
          type: string
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        reviewed_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - name
      - slug
      - created_by
      - invite_token
      - is_open
      - status
...
    FootballSeason:
      type: object
      properties:
        id:
          type: string
          format: uuid
        competition_id:
          type: string
          format: uuid
        provider:
          type: string
        provider_season_id:
          type: integer
        name:
          type: string
          nullable: true
        starts_at:
          type: string
          format: date
          nullable: true
        ends_at:
          type: string
          format: date
          nullable: true
        is_current:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FootballCompetition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        provider_league_id:
          type: integer
        name:
          type: string
        country_name:
          type: string
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        current_provider_season_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        seasons:
          type: array
          items:
            "$ref": "#/components/schemas/FootballSeason"
      required:
      - id
      - provider_league_id
      - name
    CompetitionCatalogResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/FootballCompetition"
      required:
      - data
    FriendsGroupSubscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        friends_group_id:
          type: string
          format: uuid
        competition_id:
          type: string
          format: uuid
        season_id:
          type: string
          format: uuid
          nullable: true
        provider:
          type: string
        provider_league_id:
          type: integer
        provider_season_id:
          type: integer
          nullable: true
        status:
          type: string
          enum:
          - active
          - pending
          - rejected
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HydrationResult:
      type: object
      properties:
        skipped:
          type: boolean
        reason:
          type: string
        fixturesSynced:
          type: integer
        fixturesFound:
          type: integer
        subscriptionsFound:
          type: integer
        targetsRefreshed:
          type: integer
      additionalProperties: true
    CreateFriendsGroupRequest:
      type: object
      properties:
        payload:
          type: object
          properties:
            name:
              type: string
            slug:
              type: string
              description: Lowercase kebab-case slug.
            accessType:
              type: string
              enum:
              - open
              - private
            subscription:
              type: object
              properties:
                providerLeagueId:
                  type: integer
                competitionName:
                  type: string
                  description: Optional display override. Backend prefers catalog data.
                countryName:
                  type: string
                  nullable: true
                  description: Optional display override. Backend prefers catalog data.
                logoUrl:
                  type: string
                  format: uri
                  nullable: true
                  description: Optional display override. Backend prefers catalog data.
              required:
              - providerLeagueId
          required:
          - name
          - slug
          - accessType
          - subscription
      required:
      - payload
    CreateFriendsGroupResponse:
      type: object
      properties:
        data:
          "$ref": "#/components/schemas/FriendsGroup"
        joinLink:
          type: string
        competition:
          "$ref": "#/components/schemas/FootballCompetition"
          nullable: true
        subscription:
          "$ref": "#/components/schemas/FriendsGroupSubscription"
          nullable: true
        hydration:
          "$ref": "#/components/schemas/HydrationResult"
          nullable: true
      required:
      - data
      - joinLink
    SlugCheckResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            available:
              type: boolean
              description: True when no friends group currently uses the slug.
          required:
          - available
      required:
      - data
...
    GroupCardLeague:
      type: object
      nullable: true
      properties:
        providerLeagueId:
          type: integer
        providerSeasonId:
          type: integer
          nullable: true
        name:
          type: string
        countryName:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        seasonName:
          type: string
          nullable: true
      required:
      - providerLeagueId
      - providerSeasonId
      - name
      - countryName
      - logoUrl
      - seasonName
    GroupCardMatchweek:
      type: object
      nullable: true
      properties:
        matchweek:
          type: string
          example: Matchweek 12
        weekNumber:
          type: integer
          nullable: true
          example: 12
        state:
          type: string
          enum:
          - upcoming
          - live
          - finished
        displayLabel:
          type: string
          example: Premier League week 12
      required:
      - matchweek
      - weekNumber
      - state
      - displayLabel
    GroupSeasonScoreSummary:
      type: object
      description: Current user's cumulative season/group score summary for this
        friends group. Values are totals across all scored matchweeks so far, not
        one selected matchweek.
      properties:
        user_id:
          type: string
          format: uuid
        fixtures_predicted:
          type: integer
        exact_score_points:
          type: integer
          description: Total exact scoreline bonus points across the season so far.
        correct_result_points:
          type: integer
          description: Total correct result points across the season so far.
        total_goals_bonus:
          type: integer
          description: Total matchweek total-goals bonus points across the season
            so far.
        red_card_bonus:
          type: integer
          description: Total red-card bonus points across the season so far.
        points_earned:
          type: integer
          description: Current user's season total points inside this friends group.
        weeks_played:
          type: integer
        rank:
          type: integer
          nullable: true
          description: Current user's season rank inside this friends group. Null
            when no matchweeks have been scored yet.
      required:
      - user_id
      - fixtures_predicted
      - exact_score_points
      - correct_result_points
      - total_goals_bonus
      - red_card_bonus
      - points_earned
      - weeks_played
      - rank
    MyGroup:
      type: object
      properties:
        role:
          type: string
          enum:
          - owner
          - member
        joined_at:
          type: string
          format: date-time
        friends_group:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
            created_by:
              type: string
              format: uuid
            invite_token:
              type: string
              description: Stable invite token for this friends group. FE can use it
                to rebuild the invite link after reload.
            is_open:
              type: boolean
            accessType:
              type: string
              enum:
              - open
              - private
            joinLink:
              type: string
              example: "/join/abc123"
            status:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            league:
              "$ref": "#/components/schemas/GroupCardLeague"
            currentMatchweek:
              "$ref": "#/components/schemas/GroupCardMatchweek"
          required:
          - id
          - name
          - slug
          - created_by
          - invite_token
          - is_open
          - accessType
          - joinLink
          - status
          - league
          - currentMatchweek
        score:
          "$ref": "#/components/schemas/GroupSeasonScoreSummary"
      required:
      - role
      - joined_at
      - friends_group
      - score
    MyGroupsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/MyGroup"
      required:
      - data
...